Refactor(minjae/develop): Market Component 리팩토링 진행
